Clean unused methods
