+++ src/main/java/seedu/address/model/lessons/Day.java
@@ -1,10 +1,14 @@
 package seedu.address.model.lessons;
 
-import static seedu.address.logic.parser.TypeParsingUtil.parseDate;
+import static seedu.address.logic.parser.RegularExpressionUtil.*;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
+import seedu.address.logic.parser.TypeParsingUtil;
+import seedu.address.logic.parser.exceptions.InvalidInputException;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.ListEntryField;
 
@@ -48,6 +52,72 @@ public class Day extends ListEntryField {
             return true;
         } catch (ParseException e) {
             return false;
+        }
+    }
+
+    /**
+     * Parses the date from the input string, which can be in the following formats:
+     * 1. dd, then the date will be the current month and year
+     * 2. mm/dd ,then the date will be the current year
+     * 3. yyyy/mm/dd ,then the date will be the given year
+     * 4. yy/mm/dd ,then the date will be the given year + 2000
+     * @param input the input string where the date is to be parsed from
+     * @return the date parsed
+     * @throws ParseException if the input is not a valid date
+     */
+    private static LocalDate parseDate(String input) throws ParseException {
+        String dayPattern = ONE_TO_TWO_DIGITS;
+        String monthDayPattern = ONE_TO_TWO_DIGITS + "/" + ONE_TO_TWO_DIGITS;
+        String yearMonthDayPattern2 = TWO_DIGITS + "/" + ONE_TO_TWO_DIGITS + "/" + ONE_TO_TWO_DIGITS;
+        String yearMonthDayPattern4 = FOUR_DIGITS + "/" + ONE_TO_TWO_DIGITS + "/" + ONE_TO_TWO_DIGITS;
+
+        Pattern dayP = Pattern.compile(dayPattern);
+        Pattern monthDayP = Pattern.compile(monthDayPattern);
+        Pattern yearMonthDayP2 = Pattern.compile(yearMonthDayPattern2);
+        Pattern yearMonthDayP4 = Pattern.compile(yearMonthDayPattern4);
+
+        Matcher dayM = dayP.matcher(input);
+        Matcher monthDayM = monthDayP.matcher(input);
+        Matcher yearMonthDayM2 = yearMonthDayP2.matcher(input);
+        Matcher yearMonthDayM4 = yearMonthDayP4.matcher(input);
+
+        if (yearMonthDayM2.matches()) {
+            int year = TypeParsingUtil.parseNum(yearMonthDayM2.group(1), 0, 99) + 2000;
+            int month = TypeParsingUtil.parseNum(yearMonthDayM2.group(2), 1, 12);
+            int day = TypeParsingUtil.parseNum(yearMonthDayM2.group(3), 1, findMaxDay(year, month));
+            return LocalDate.of(year, month, day);
+        } else if (yearMonthDayM4.matches()) {
+            int year = TypeParsingUtil.parseNum(yearMonthDayM4.group(1), 0, 9999);
+            int month = TypeParsingUtil.parseNum(yearMonthDayM4.group(2), 1, 12);
+            int day = TypeParsingUtil.parseNum(yearMonthDayM4.group(3), 1, findMaxDay(year, month));
+            return LocalDate.of(year, month, day);
+        } else if (monthDayM.matches()) {
+            LocalDate now = LocalDate.now();
+            int month = TypeParsingUtil.parseNum(monthDayM.group(1), 1, 12);
+            int day = TypeParsingUtil.parseNum(monthDayM.group(2), 1, findMaxDay(now.getYear(), month));
+            return LocalDate.of(now.getYear(), month, day);
+        } else if (dayM.matches()) {
+            LocalDate now = LocalDate.now();
+            int day = TypeParsingUtil.parseNum(dayM.group(1), 1, findMaxDay(now.getYear(), now.getMonthValue()));
+            return LocalDate.of(now.getYear(), now.getMonth(), day);
+        } else {
+            throw new InvalidInputException(input
+                    + " is not a valid date, please use yyyy/mm/dd or mm/dd or dd"
+                    + "\nfor example, assume today is 2023/11/3, to add 2023/11/29, could"
+                    + " use 29, 11/29, 2023/11/29 or 23/11/29");
+        }
+    }
+    private static Integer findMaxDay(int year, int month) {
+        if (month == 2) {
+            if (year % 4 == 0) {
+                return 29;
+            } else {
+                return 28;
+            }
+        } else if (month == 4 || month == 6 || month == 9 || month == 11) {
+            return 30;
+        } else {
+            return 31;
         }
     }
     @Override
