+++ src/test/java/seedu/address/logic/parser/AddTaskCommandParserTest.java
@@ -1,0 +1,9 @@
+package seedu.address.logic.parser;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+// partition based on if the input has valid index and valid description.
+// addtask with index, without index when no current lesson is shown , without index when index is invalid
+class AddTaskCommandParserTest {
+
+}
